+++ frontend/src/Frontend/Crypto/Browser.hs
@@ -50,7 +50,6 @@ instance (MonadJSM m, MonadSample t m) => HasCrypto PrivateKey (BrowserCryptoT t
       --TODO FIX THIS
       Nothing -> undefined
       Just pair -> pure pair
-
   cryptoGenPubKeyFromPrivate pkScheme t = do
     case textToKey t of
       Nothing -> pure $ Left $ "cryptoGenPubKeyFromPrivate: not a valid private key"
