+++ frontend/src/Frontend/Crypto/Ed25519.hs
@@ -39,6 +39,7 @@ module Frontend.Crypto.Ed25519
   ----------------------
   , generateRoot
   , generateKeypair
+  , genMnemonic
   )
   where
