+++ frontend/src/Frontend/Setup/Common.hs
@@ -9,7 +9,11 @@
 {-# LANGUAGE TypeFamilies #-}
 
 -- | Wallet setup screens
-module Frontend.GHCOnly.Setup (WalletExists (..), Password(..), runSetup, splashLogo, setupDiv, setupClass, checkPassword) where
+module Frontend.Setup.Common 
+  ( 
+    WalletExists (..), Password(..), runSetup
+  , splashLogo, setupDiv, setupClass, checkPassword
+  ) where
 
 import Control.Lens ((<>~), (??), (^.), _1, _2, _3)
 import Control.Error (hush)
@@ -29,27 +33,19 @@ import Data.Functor ((<&>))
 import Data.Text (Text)
 import Language.Javascript.JSaddle (MonadJSM, liftJSM)
 import Reflex.Dom.Core
--- import qualified Cardano.Crypto.Wallet as Crypto
--- import qualified Crypto.Encoding.BIP39 as Crypto
--- import qualified Crypto.Encoding.BIP39.English as Crypto
--- import qualified Crypto.Random.Entropy
 import qualified Data.ByteArray as BA
 import qualified Data.Map as Map
 import qualified Data.Text as T
 import qualified Data.Text.Encoding as T
-import System.FilePath (takeFileName)
 
-import Frontend.AppCfg (FileFFI(..), FileType(FileType_Import))
--- import Desktop.ImportExport (doImport, Password(..), ImportWalletError(..))
--- import Pact.Server.ApiClient (HasTransactionLogger, askTransactionLogger)
+import Frontend.AppCfg (FileFFI(..))
+import Frontend.Crypto.Ed25519
 import Frontend.Storage.Class (HasStorage)
 import Frontend.UI.Button
 import Frontend.UI.Dialogs.ChangePassword (minPasswordLength)
 import Frontend.UI.Widgets.Helpers (imgWithAlt)
 import Frontend.UI.Widgets
 import Obelisk.Generated.Static
-
-import Frontend.Crypto.Ed25519
 
 newtype Password = Password { unPassword :: Text } deriving (Eq)
 
@@ -96,23 +92,15 @@ setupCheckbox
 setupCheckbox initialValue cfg inner = elClass "div" (setupClass "checkbox")
   $ uiCheckbox def initialValue cfg inner
 
--- mkPhraseMapFromMnemonic
---   :: forall mw.
---      Crypto.ValidMnemonicSentence mw
---   => Crypto.MnemonicSentence mw
---   -> Map.Map WordKey Text
--- mkPhraseMapFromMnemonic = wordsToPhraseMap . T.words . baToText
---   . Crypto.mnemonicSentenceToString @mw Crypto.english
-
 showWordKey :: WordKey -> Text
 showWordKey = T.pack . show . _unWordKey
 
 -- | Convenience function for unpacking byte array things into 'Text'
-baToText :: ByteArrayAccess b => b -> Text
-baToText = T.decodeUtf8 . BA.pack . BA.unpack
+-- baToText :: ByteArrayAccess b => b -> Text
+-- baToText = T.decodeUtf8 . BA.pack . BA.unpack
 
-textTo :: IsString a => Text -> a
-textTo = fromString . T.unpack
+-- textTo :: IsString a => Text -> a
+-- textTo = fromString . T.unpack
 
 tshow :: Show a => a -> Text
 tshow = T.pack . show
@@ -248,7 +236,6 @@ splashScreen
   :: (DomBuilder t m, MonadFix m, MonadHold t m, PerformEvent t m
      , PostBuild t m, MonadJSM (Performable m), TriggerEvent t m, HasStorage (Performable m)
      , MonadSample t (Performable m)
-     -- , HasTransactionLogger m
      )
   => WalletExists
   -> FileFFI t m
@@ -281,11 +268,9 @@ splashScreen walletExists fileFFI eBack = selfWF
           $ text "Restore from wallet export"
 
         finishSetupWF WalletScreen_SplashScreen $ leftmost
-          [ restoreBipWallet selfWF eBack <$ hasAgreed restoreBipPhrase]
-          -- [ createNewWallet selfWF eBack <$ hasAgreed create
-          -- , restoreBipWallet selfWF eBack <$ hasAgreed restoreBipPhrase
-          -- , restoreFromImport walletExists fileFFI selfWF eBack <$ hasAgreed restoreImport
-          -- ]
+          [ createNewWallet selfWF eBack <$ hasAgreed create
+          , restoreBipWallet selfWF eBack <$ hasAgreed restoreBipPhrase
+          ]
 
 data BIP39PhraseError
   = BIP39PhraseError_InvalidPhrase
@@ -293,9 +278,6 @@ data BIP39PhraseError
 
 passphraseLen :: Int
 passphraseLen = 12
-
--- sentenceToSeed :: Crypto.ValidMnemonicSentence mw => Crypto.MnemonicSentence mw -> Crypto.Seed
--- sentenceToSeed s = Crypto.sentenceToSeed s Crypto.english ""
 
 restoreBipWallet
   :: (DomBuilder t m, MonadHold t m, MonadFix m, PerformEvent t m, PostBuild t m, 
@@ -332,7 +314,6 @@ restoreBipWallet backWF eBack = Workflow $ do
         -> mempty
 
   -- GEN SEED: Todo: Make sure wrong words returns Nothing
-      -- generateRoot $ T.unwords phrase'
   let 
       ePhraseUpdated = fmapMaybe hush $ updated sentenceOrError
 
@@ -340,7 +321,6 @@ restoreBipWallet backWF eBack = Workflow $ do
         text "Waiting for a valid 12 word passphrase..."
         pure never
 
-      -- withSeedConfirmPassword ::_
       withSeedConfirmPassword sentence = setupDiv "recover-enter-password" $ mdo
         (ePwSubmit, dSetPw) <- restoreForm (fmap isNothing dSetPw) $ do
           holdDyn Nothing =<< setPassword (pure sentence)
@@ -352,103 +332,6 @@ restoreBipWallet backWF eBack = Workflow $ do
     ( (WalletScreen_RecoverPassphrase, (\(prv, pw) -> (prv, pw, True)) <$> switchDyn dSetPassword, never)
     , backWF <$ eBack
     )
-
--- restoreFromImport
---   :: forall t m
---   .  ( DomBuilder t m, MonadFix m, MonadHold t m, PerformEvent t m
---      , PostBuild t m, MonadJSM (Performable m), HasStorage (Performable m)
---      , MonadSample t (Performable m)
---      , HasTransactionLogger m
---      )
---   => WalletExists
---   -> FileFFI t m
---   -> SetupWF t m
---   -> Event t ()
---   -> SetupWF t m
--- restoreFromImport walletExists fileFFI backWF eBack = nagScreen
---   where
---     nagMsgs = case walletExists of
---       WalletExists_Yes ->
---         ("You are about to replace the current wallet's data"
---         ,"Reminder: Importing a wallet file will replace the data within the current wallet."
---         )
---       WalletExists_No ->
---         ("Please select the wallet import file."
---         ,"Reminder: You will need your wallet password to proceed."
---         )
-
---     nagBack = case walletExists of
---       WalletExists_No -> pure never
---       WalletExists_Yes -> uiButtonDyn
---         -- TODO: Don't reuse this class or at least rename it
---         (btnCfgSecondary & uiButtonCfg_class <>~ "setup__restore-existing-button")
---         (text "Go back and export current wallet")
-
---     nagScreen = Workflow $ setupDiv "splash" $ do
---       splashLogo
---       let (nagTitle, nagReminder) = nagMsgs
---       elClass "h1" "setup__recover-import-title" $ text nagTitle
---       elClass "p" "setup__recover-import-text" $ text nagReminder
---       eImport <- confirmButton def "Select Import File"
---       eExit <- nagBack
---       pure
---         ( (WalletScreen_RecoverImport, never, eExit)
---         , leftmost
---           [ backWF <$ (eBack <> eExit)
---           , importScreen <$ eImport
---           ]
---         )
-
---     importScreen = Workflow $ setupDiv "splash" $ mdo
---       splashLogo
---       elClass "h1" "setup__recover-import-title" $ text "Import File Password"
---       elClass "p" "setup__recover-import-text" $ text "Enter the password for the chosen wallet file in order to authorize access to the data."
-
---       let disabled = isNothing <$> dmValidForm
---       dErr <- holdDyn Nothing (leftmost [Just <$> eImportErr, Nothing <$ updated dmValidForm])
---       (eSubmit, (dFileSelected, pwInput)) <- setupForm "" "Import File" disabled $ mdo
---         ePb <- getPostBuild
---         (selectElt, _) <- elClass' "div" "setup__recover-import-file" $ do
---           imgWithAlt (static @"img/import.svg") "Import" blank
---           divClass "setup__recover-import-file-text" $ dynText $ ffor dFileSelected $
---             maybe "Select a file" (T.pack . takeFileName . fst)
-
---         performEvent_ $ liftJSM (_fileFFI_openFileDialog fileFFI FileType_Import) <$
---           ((domEvent Click selectElt) <> ePb)
-
---         dFileSelected <- holdDyn Nothing (Just <$> _fileFFI_externalFileOpened fileFFI)
-
---         pw <- uiPassword (setupClass "password-wrapper") (setupClass "password") "Enter import wallet password"
-
---         dyn_ $ ffor dErr $ traverse_ $ \err ->
---           elClass "p" "error_inline" $ text $ case err of
---             ImportWalletError_InvalidCommandLogDestination -> "Destination for transaction log file is invalid"
---             ImportWalletError_CommandLogWriteError -> "Unable to write transaction log file"
---             ImportWalletError_PasswordIncorrect -> "Incorrect Password"
---             ImportWalletError_NoRootKey -> "Backup cannot be restored as it does not contain a BIP Root Key"
---             ImportWalletError_NotJson eMsg -> "Backup cannot be restored as it is not a valid json file. Error: " <> eMsg
---             ImportWalletError_DecodeError section ver eMsg ->
---               "Backup section " <> section <> " cannot be parsed as version " <> tshow ver  <>  " with error: " <> eMsg
---             ImportWalletError_UnknownVersion section ver ->
---               "Backup section " <> section <> " has an unknown version " <> tshow ver <> ". It's likely that this backup is from a newer version of chainweaver."
-
-
---         pure (dFileSelected, pw)
-
---       eExit <- nagBack
---       let dmValidForm = runMaybeT $ (,)
---             <$> MaybeT (nonEmptyPassword <$> (_inputElement_value pwInput))
---             <*> MaybeT (fmap snd <$> dFileSelected)
-
---       txLogger <- askTransactionLogger
---       eImport <- performEvent $ tagMaybe (fmap (uncurry (doImport @t txLogger)) <$> current dmValidForm) eSubmit
-
---       let (eImportErr, eImportDone) = fanEither eImport
-
---       pure
---         ( (WalletScreen_RecoverImport, (\(prv,pw) -> (prv, pw, False)) <$> eImportDone, eExit)
---         , backWF <$ (eBack <> eExit)
---         )
 
 --     nonEmptyPassword "" = Nothing
 --     nonEmptyPassword pw = Just (Password pw)
@@ -514,41 +397,42 @@ passphraseWidget dWords dStage exposeEmpty = do
   setupDiv "passphrase-widget-wrapper" $
     listViewWithKey dWords (passphraseWordElement dStage exposeEmpty)
 
--- createNewWallet
---   :: forall t m. (DomBuilder t m, MonadFix m, MonadHold t m, PerformEvent t m, PostBuild t m, MonadJSM (Performable m), TriggerEvent t m)
---   => SetupWF t m -> Event t () -> SetupWF t m
--- createNewWallet backWF eBack = selfWF
---   where
---     selfWF = Workflow $  do
---       ePb <- getPostBuild
---       elAttr "img" ("src" =: static @"img/Wallet_Graphic_2.png" <> "class" =: setupClass "password-bg") blank
+createNewWallet
+  :: forall t m. (DomBuilder t m, MonadFix m, MonadHold t m, PerformEvent t m, PostBuild t m,
+                  MonadJSM (Performable m), MonadSample t (Performable m), TriggerEvent t m)
+  => SetupWF t m -> Event t () -> SetupWF t m
+createNewWallet backWF eBack = selfWF
+  where
+    selfWF = Workflow $  do
+      ePb <- getPostBuild
+      elAttr "img" ("src" =: static @"img/Wallet_Graphic_2.png" <> "class" =: setupClass "password-bg") blank
 
---       el "h1" $ text "Set a password"
---       setupDiv "new-wallet-password-text" $ do
---         el "div" $ text "Enter a strong and unique password"
---         el "div" $ text "to protect access to your Chainweaver wallet"
+      el "h1" $ text "Set a password"
+      setupDiv "new-wallet-password-text" $ do
+        el "div" $ text "Enter a strong and unique password"
+        el "div" $ text "to protect access to your Chainweaver wallet"
 
---       (eGenError, eGenSuccess) <- fmap fanEither . performEvent $ genMnemonic <$ ePb
+      eGen <- performEvent $ liftJSM genMnemonic <$ ePb
 
---       let
---         generating = do
---           dynText =<< holdDyn "Generating your mnemonic..." eGenError
---           pure never
+      let
+        generating = do
+          text "Generating your mnemonic"
+          pure never
 
---         proceed :: Crypto.MnemonicSentence 12 -> m (Event t (SetupWF t m))
---         proceed mnem = mdo
---           (ePwSubmit, dPassword) <- continueForm (fmap isNothing dPassword) $ do
---             holdDyn Nothing =<< setPassword (pure $ sentenceToSeed mnem)
+        proceed :: [Text] -> m (Event t (SetupWF t m))
+        proceed mnem = mdo
+          (ePwSubmit, dPassword) <- continueForm (fmap isNothing dPassword) $ do
+            holdDyn Nothing =<< setPassword (pure mnem)
 
---           fmap snd $ runWithReplace blank $ ffor (tagMaybe (current dPassword) ePwSubmit) $ \p ->
---             pure $ precreatePassphraseWarning selfWF eBack p mnem
+          fmap snd $ runWithReplace blank $ ffor (tagMaybe (current dPassword) ePwSubmit) $ \p ->
+            pure $ precreatePassphraseWarning selfWF eBack p mnem
 
---       dContinue <- widgetHold generating (proceed <$> eGenSuccess)
+      dContinue <- widgetHold generating (proceed <$> eGen)
 
---       finishSetupWF WalletScreen_Password $ leftmost
---         [ backWF <$ eBack
---         , switchDyn dContinue
---         ]
+      finishSetupWF WalletScreen_Password $ leftmost
+        [ backWF <$ eBack
+        , switchDyn dContinue
+        ]
 
 walletSplashWithIcon :: DomBuilder t m => m () -> m ()
 walletSplashWithIcon w = do
@@ -567,39 +451,39 @@ stackFaIcon icon = elClass "span" "fa-stack fa-5x" $ do
   elClass "i" "fa fa-circle fa-stack-2x" blank
   elClass "i" ("fa " <> icon <> " fa-stack-1x fa-inverse") blank
 
--- precreatePassphraseWarning
---   :: (DomBuilder t m, MonadFix m, MonadHold t m, PerformEvent t m, PostBuild t m, MonadJSM (Performable m))
---   => SetupWF t m
---   -> Event t ()
---   -> (Crypto.XPrv, Password)
---   -> Crypto.MnemonicSentence 12
---   -> SetupWF t m
--- precreatePassphraseWarning backWF eBack (rootKey, password) mnemonicSentence = Workflow $ mdo
---   (eContinue, dUnderstand) <- continueForm (fmap not dUnderstand) $ do
+precreatePassphraseWarning
+  :: (DomBuilder t m, MonadFix m, MonadHold t m, PerformEvent t m, PostBuild t m, MonadJSM (Performable m))
+  => SetupWF t m
+  -> Event t ()
+  -> (PrivateKey, Password)
+  -> [Text]
+  -> SetupWF t m
+precreatePassphraseWarning backWF eBack (rootKey, password) mnemonicSentence = Workflow $ mdo
+  (eContinue, dUnderstand) <- continueForm (fmap not dUnderstand) $ do
 
---     setupDiv "warning-splash" $ do
---       walletSplashWithIcon $ do
---         setupDiv "repeat-icon" $ stackFaIcon "fa-repeat"
+    setupDiv "warning-splash" $ do
+      walletSplashWithIcon $ do
+        setupDiv "repeat-icon" $ stackFaIcon "fa-repeat"
 
---     el "h1" $ text "Wallet Recovery Phrase"
+    el "h1" $ text "Wallet Recovery Phrase"
 
---     setupDiv "recovery-phrase-warning" $ do
---       line "In the next step you will record your 12 word recovery phrase."
---       line "Your recovery phrase makes it easy to restore your wallet on a new device."
---       line "Anyone with this phrase can take control of your wallet, keep this phrase private."
---       line "Kadena cannot access your recovery phrase if lost, please store it safely."
+    setupDiv "recovery-phrase-warning" $ do
+      line "In the next step you will record your 12 word recovery phrase."
+      line "Your recovery phrase makes it easy to restore your wallet on a new device."
+      line "Anyone with this phrase can take control of your wallet, keep this phrase private."
+      line "Kadena cannot access your recovery phrase if lost, please store it safely."
 
---     let chkboxcls = setupClass "checkbox-wrapper"
---     fmap value $ elClass "div" chkboxcls $ setupCheckbox False def $ el "div" $ do
---       line "I understand that if I lose my recovery phrase,"
---       line "I will not be able to restore my wallet."
+    let chkboxcls = setupClass "checkbox-wrapper"
+    fmap value $ elClass "div" chkboxcls $ setupCheckbox False def $ el "div" $ do
+      line "I understand that if I lose my recovery phrase,"
+      line "I will not be able to restore my wallet."
 
---   finishSetupWF WalletScreen_PrePassphrase $ leftmost
---     [ backWF <$ eBack
---     , createNewPassphrase backWF eBack (rootKey, password) mnemonicSentence <$ eContinue
---     ]
---   where
---     line = el "div" . text
+  finishSetupWF WalletScreen_PrePassphrase $ leftmost
+    [ backWF <$ eBack
+    , createNewPassphrase backWF eBack (rootKey, password) mnemonicSentence <$ eContinue
+    ]
+  where
+    line = el "div" . text
 
 doneScreen
   :: (DomBuilder t m, PostBuild t m)
@@ -618,72 +502,73 @@ doneScreen (rootKey, passwd) = Workflow $ do
        )
 
 -- | UI for generating and displaying a new mnemonic sentence.
--- createNewPassphrase
---   :: (DomBuilder t m, MonadFix m, MonadHold t m, PerformEvent t m, PostBuild t m, MonadJSM (Performable m))
---   => SetupWF t m
---   -> Event t ()
---   -> (Crypto.XPrv, Password)
---   -> Crypto.MnemonicSentence 12
---   -> SetupWF t m
--- createNewPassphrase backWF eBack (rootKey, password) mnemonicSentence = selfWF
---   where
---     selfWF = Workflow $ mdo
---       (eContinue, dIsStored) <- continueForm (fmap not dIsStored) $ do
---         el "h1" $ text "Record Recovery Phrase"
---         el "div" $ do
---           el "div" $ text "Write down or copy these words in the correct order and store them safely."
---           el "div" $ text "The recovery words are hidden for security. Mouseover the numbers to reveal."
+createNewPassphrase
+  :: (DomBuilder t m, MonadFix m, MonadHold t m, PerformEvent t m, PostBuild t m, MonadJSM (Performable m))
+  => SetupWF t m
+  -> Event t ()
+  -> (PrivateKey, Password)
+  -- TODO: Wrap Sentece type
+  -> [ Text ]
+  -> SetupWF t m
+createNewPassphrase backWF eBack (rootKey, password) mnemonicSentence = selfWF
+  where
+    selfWF = Workflow $ mdo
+      (eContinue, dIsStored) <- continueForm (fmap not dIsStored) $ do
+        el "h1" $ text "Record Recovery Phrase"
+        el "div" $ do
+          el "div" $ text "Write down or copy these words in the correct order and store them safely."
+          el "div" $ text "The recovery words are hidden for security. Mouseover the numbers to reveal."
 
---         rec
---           dPassphrase <- passphraseWidget dPassphrase (pure Setup) False
---             >>= holdDyn (mkPhraseMapFromMnemonic mnemonicSentence)
+        rec
+          dPassphrase <- passphraseWidget dPassphrase (pure Setup) False
+            >>= holdDyn (wordsToPhraseMap mnemonicSentence)
 
---           let cfg = def
---                 & uiButtonCfg_class .~ "setup__recovery-phrase-copy"
---                 & uiButtonCfg_title .~ pure (Just "Copy")
---           copyButton cfg True $
---             T.unwords . Map.elems <$> current dPassphrase
+          let cfg = def
+                & uiButtonCfg_class .~ "setup__recovery-phrase-copy"
+                & uiButtonCfg_title .~ pure (Just "Copy")
+          copyButton cfg True $
+            T.unwords . Map.elems <$> current dPassphrase
 
---         fmap value $ setupDiv "checkbox-wrapper" $ setupCheckbox False def
---           $ text "I have safely stored my recovery phrase."
+        fmap value $ setupDiv "checkbox-wrapper" $ setupCheckbox False def
+          $ text "I have safely stored my recovery phrase."
 
---       finishSetupWF WalletScreen_CreatePassphrase $ leftmost
---         [ backWF <$ eBack
---         , confirmPhrase selfWF eBack (rootKey, password) mnemonicSentence <$ eContinue
---         ]
+      finishSetupWF WalletScreen_CreatePassphrase $ leftmost
+        [ backWF <$ eBack
+        , confirmPhrase selfWF eBack (rootKey, password) mnemonicSentence <$ eContinue
+        ]
 
 -- | UI for mnemonic sentence confirmation: scramble the phrase, make the user
 -- choose the words in the right order.
--- confirmPhrase
---   :: (DomBuilder t m, MonadFix m, MonadHold t m, PostBuild t m)
---   => SetupWF t m
---   -> Event t ()
---   -> (Crypto.XPrv, Password)
---   -> Crypto.MnemonicSentence 12
---   -- ^ Mnemonic sentence to confirm
---   -> SetupWF t m
--- confirmPhrase backWF eBack (rootKey, password) mnemonicSentence = Workflow $ mdo
---   (continue, done) <- continueForm (fmap not done) $ do
---     el "h1" $ text "Verify Recovery Phrase"
---     el "div" $ do
---       el "div" $ text "Please confirm your recovery phrase by"
---       el "div" $ text "typing the words in the correct order."
+confirmPhrase
+  :: (DomBuilder t m, MonadFix m, MonadHold t m, PostBuild t m)
+  => SetupWF t m
+  -> Event t ()
+  -> (PrivateKey, Password)
+  -> [Text]
+  -- ^ Mnemonic sentence to confirm
+  -> SetupWF t m
+confirmPhrase backWF eBack (rootKey, password) mnemonicSentence = Workflow $ mdo
+  (continue, done) <- continueForm (fmap not done) $ do
+    el "h1" $ text "Verify Recovery Phrase"
+    el "div" $ do
+      el "div" $ text "Please confirm your recovery phrase by"
+      el "div" $ text "typing the words in the correct order."
 
---     let actualMap = mkPhraseMapFromMnemonic mnemonicSentence
+    let actualMap = wordsToPhraseMap mnemonicSentence
 
---     rec
---       onPhraseUpdate <- divClass "setup__verify-passphrase-wrapper" $
---         passphraseWidget dConfirmPhrase (pure Recover) True
+    rec
+      onPhraseUpdate <- divClass "setup__verify-passphrase-wrapper" $
+        passphraseWidget dConfirmPhrase (pure Recover) True
 
---       dConfirmPhrase <- holdDyn (wordsToPhraseMap $ replicate passphraseLen T.empty)
---         $ flip Map.union <$> current dConfirmPhrase <@> onPhraseUpdate
+      dConfirmPhrase <- holdDyn (wordsToPhraseMap $ replicate passphraseLen T.empty)
+        $ flip Map.union <$> current dConfirmPhrase <@> onPhraseUpdate
 
---     pure $ (== actualMap) <$> dConfirmPhrase
+    pure $ (== actualMap) <$> dConfirmPhrase
 
---   finishSetupWF WalletScreen_VerifyPassphrase $ leftmost
---     [ doneScreen (rootKey, password) <$ continue
---     , backWF <$ eBack
---     ]
+  finishSetupWF WalletScreen_VerifyPassphrase $ leftmost
+    [ doneScreen (rootKey, password) <$ continue
+    , backWF <$ eBack
+    ]
 
 setPassword
   :: (DomBuilder t m, MonadHold t m, MonadFix m, PerformEvent t m, PostBuild t m, 
@@ -721,9 +606,6 @@ setPassword dSentence = do
   elDynClass "div" dMsgClass $
     dynText $ fromMaybe T.empty <$> lastError
 
-  --Web version doesn't actually encrypt the root
-  -- eRoot <- performEvent (updated dSentence) $ (liftJSM . generateRoot . T.unwords)
-
   encryptedKeyAndPass <- performEvent $ ffor pass $ \p -> do
     sentence <- sample $ current dSentence 
     root <- liftJSM $ generateRoot $ T.unwords sentence
@@ -743,12 +625,3 @@ checkPassword p1 p2
       Left "Passwords must match"
   | otherwise =
       Right p1
-
--- | Generate a 12 word mnemonic sentence, using cryptonite.
---
--- These values for entropy must be set according to a predefined table:
--- https://github.com/kadena-io/cardano-crypto/blob/master/src/Crypto/Encoding/BIP39.hs#L208-L218
--- genMnemonic :: MonadIO m => m (Either Text (Crypto.MnemonicSentence 12))
--- genMnemonic = liftIO $ bimap tshow Crypto.entropyToWords . Crypto.toEntropy @128
---   -- This size must be a 1/8th the size of the 'toEntropy' size: 128 / 8 = 16
---   <$> Crypto.Random.Entropy.getEntropy @ByteString 16
